# vpn-gate-backend/docs/swagger.yaml
basePath: /
definitions:
  vpn-gate-backend_internal_database.VpnServer:
    properties:
      countryLong:
        example: Japan
        type: string
      countryShort:
        example: JP
        type: string
      hostName:
        example: vpn123456789.opengw.net
        type: string
      ip:
        example: 1.2.3.4
        type: string
      isActive:
        example: true
        type: boolean
      lastScraped:
        example: "2026-01-15T10:30:00Z"
        type: string
      lastSeen:
        example: "2026-01-15T10:30:00Z"
        type: string
      method:
        enum:
        - UDP
        - TCP
        example: UDP
        type: string
      openVpnConfigBase64:
        example: IyEvdXNyL2Jpbi9lbnYgYmFzaA==
        type: string
      operator:
        example: VPNGate
        type: string
      ping:
        example: 42
        type: integer
      port:
        example: 1194
        type: integer
      score:
        example: 95000
        type: integer
      serverType:
        enum:
        - RESIDENTIAL
        - ACADEMIC
        - DATACENTER
        example: DATACENTER
        type: string
      source:
        example: vpngate
        type: string
      speed:
        example: 102400000
        type: integer
      uptime:
        example: 30d
        type: string
      vpnChecked:
        example: true
        type: boolean
      vpnDetected:
        example: false
        type: boolean
    type: object
info:
  contact:
    name: Zenith VPN
  description: Backend API for Zenith VPN client — scrapes, validates, and serves
    VPN Gate server lists.
  license:
    name: MIT
  title: Zenith VPN Gate API
  version: "1.0"
paths:
  /api/health:
    get:
      description: Returns service health status including database connectivity.
      produces:
      - application/json
      responses:
        "200":
          description: healthy
          schema:
            additionalProperties: true
            type: object
        "503":
          description: degraded
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - system
  /api/servers:
    get:
      consumes:
      - application/json
      description: Returns all currently active (online) VPN servers. Optionally filter
        by server classification type and/or country.
      parameters:
      - description: Server classification filter
        enum:
        - RESIDENTIAL
        - ACADEMIC
        - DATACENTER
        in: query
        name: type
        type: string
      - description: Country short code filter (e.g. JP, US, DE)
        example: '"JP"'
        in: query
        name: country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vpn-gate-backend_internal_database.VpnServer'
            type: array
        "400":
          description: Invalid server type
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List active VPN servers
      tags:
      - servers
  /api/servers/all:
    get:
      consumes:
      - application/json
      description: Returns every known VPN server regardless of active status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vpn-gate-backend_internal_database.VpnServer'
            type: array
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List all VPN servers
      tags:
      - servers
  /api/servers/ip/{ip}:
    get:
      consumes:
      - application/json
      description: Returns full server details including OpenVPN config for connecting.
      parameters:
      - description: Server IP address
        in: path
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vpn-gate-backend_internal_database.VpnServer'
        "404":
          description: Server not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get server by IP
      tags:
      - servers
produces:
- application/json
swagger: "2.0"